Logs and exception handler added
